chore: cleanup organization of the process managers
They were all over the place, without any naming consistency.
Move them all to syskit/process_managers and call the manager
Manager and the process Process.
