+++ lib/syskit/roby_app.rb
@@ -17,18 +17,12 @@ require "syskit/roby_app/logging_configuration"
 require "syskit/roby_app/logging_group"
 require "syskit/roby_app/robot_extension"
 require "syskit/roby_app/toplevel"
-require "syskit/roby_app/log_transfer_manager"
 require "syskit/roby_app/configuration"
 require "syskit/roby_app/plugin"
 require "syskit/roby_app/single_file_dsl"
-require "syskit/roby_app/unmanaged_process"
-require "syskit/roby_app/unmanaged_tasks_manager"
-require "syskit/roby_app/in_process_deployment"
-require "syskit/roby_app/in_process_tasks_manager"
-require "syskit/roby_app/ruby_tasks/process"
-require "syskit/roby_app/ruby_tasks/process_manager"
-require "syskit/roby_app/remote_processes"
+
 require "syskit/roby_app/tmp_root_ca"
+require "syskit/roby_app/log_transfer_manager"
 require "syskit/roby_app/log_transfer_server"
 
 module Syskit
