+++ test/roby_app/test_configuration.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "syskit/test/self"
-require "syskit/roby_app/remote_processes"
 
 describe Syskit::RobyApp::Configuration do
     describe "#use_deployment" do
@@ -9,8 +8,12 @@ describe Syskit::RobyApp::Configuration do
         before do
             @task_m = Syskit::TaskContext.new_submodel(name: "test::Task")
             @conf = Syskit::RobyApp::Configuration.new(Roby.app)
-            conf.register_process_server("localhost", Syskit::RobyApp::RubyTasks::ProcessManager.new(Roby.app.default_loader), "")
-            conf.register_process_server("test", Syskit::RobyApp::RubyTasks::ProcessManager.new(Roby.app.default_loader), "")
+            conf.register_process_server(
+                "localhost", Syskit::ProcessManagers::RubyTasks::Manager.new
+            )
+            conf.register_process_server(
+                "test", Syskit::ProcessManagers::RubyTasks::Manager.new
+            )
         end
 
         def stub_deployment(name)
@@ -39,8 +42,9 @@ describe Syskit::RobyApp::Configuration do
         before do
             @task_m = Syskit::RubyTaskContext.new_submodel
             @conf = Syskit::RobyApp::Configuration.new(Roby.app)
-            @conf.register_process_server("ruby_tasks",
-                                          Syskit::RobyApp::RubyTasks::ProcessManager.new(Roby.app.default_loader), "")
+            @conf.register_process_server(
+                "ruby_tasks", Syskit::ProcessManagers::RubyTasks::Manager.new
+            )
         end
 
         it "defines a deployment for a given ruby task context model" do
@@ -55,8 +59,9 @@ describe Syskit::RobyApp::Configuration do
         before do
             @task_m = Syskit::TaskContext.new_submodel
             @conf = Syskit::RobyApp::Configuration.new(Roby.app)
-            @conf.register_process_server("unmanaged_tasks",
-                                          Syskit::RobyApp::RubyTasks::ProcessManager.new(Roby.app.default_loader), "")
+            @conf.register_process_server(
+                "unmanaged_tasks", Syskit::ProcessManagers::RubyTasks::Manager.new
+            )
         end
         it "defines a configured deployment from a task model and name" do
             configured_deployments = @conf.use_unmanaged_task @task_m => "name"
@@ -81,7 +86,7 @@ describe Syskit::RobyApp::Configuration do
             @available_deployment_names = ["deployment"]
 
             set_log_level ::Robot, Logger::FATAL + 1
-            set_log_level Syskit::RobyApp::RemoteProcesses::Server, Logger::FATAL + 1
+            set_log_level Syskit::ProcessManagers::Remote::Server, Logger::FATAL + 1
         end
 
         after do
@@ -142,7 +147,7 @@ describe Syskit::RobyApp::Configuration do
             @server_loader.should_receive(:each_available_project_name)
                           .and_return { @available_project_names }
 
-            @server = Syskit::RobyApp::RemoteProcesses::Server.new(
+            @server = Syskit::ProcessManagers::Remote::Server::Server.new(
                 Roby.app, port: 0, loader: @server_loader
             )
             flexmock(@server)
